refactor(api): separate out fungibles errors module
